Update contracts and error messages
