+++ modules/meeting/app/contracts/meeting_outcomes/base_contract.rb
@@ -30,7 +30,7 @@
 
 module MeetingOutcomes
   class BaseContract < ::ModelContract
-    include ModifiableItem
+    include EditableItem
 
     def self.model
       MeetingOutcome
