+++ modules/meeting/app/models/meeting_outcome.rb
@@ -41,14 +41,6 @@ class MeetingOutcome < ApplicationRecord
   validates_presence_of :notes, if: -> { information_kind? }
   validates_presence_of :work_package, if: -> { work_package_kind? }
 
-  # validate :unique_kind
-
-  # def unique_kind
-  #   if meeting_agenda_item.outcomes.exists?(kind: kind)
-  #     errors.add(:kind, "has already been taken for this MeetingAgenda")
-  #   end
-  # end
-
   def editable?
     meeting_agenda_item.meeting&.in_progress?
   end
